CWS: Mark ajax request endpoints as @api_login_required
This prevents cases where they would redirect to the login screen
unexpectedly. Also improved handling of errors in the submission details
popup.
