+++ cms/server/contest/templates/test_interface.html
@@ -25,7 +25,11 @@ $(document).on("click", ".user_test_list tbody tr td.status .details", function 
     var modal = $("#user_test_detail");
     var modal_body = modal.children(".modal-body");
     modal_body.html('<div class="loading"><img src="{{ url("static", "loading.gif") }}"/>{% trans %}loading...{% endtrans %}</div>');
-    modal_body.load(utils.contest_url("tasks", task_id, "tests", user_test_id, "details"));
+    modal_body.load(utils.contest_url("tasks", task_id, "tests", user_test_id, "details"), function(response, status, xhr) {
+        if(status != "success") {
+            $(this).html("{% trans %}Error loading details, please refresh the page.{% endtrans %}");
+        }
+    });
     modal.modal("show");
 });
